finish product api

+++ E-Commerce.Web/Areas/Api/Controllers/ProductController.cs
@@ -2,14 +2,15 @@
 using E_Commerce.Business.Services.Interfaces;
 using E_Commerce.DataAccess.Repositories.Interfaces;
 using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using E_Commerce.Business.Services.Implementation;
+using Microsoft.AspNetCore.Authorization;
+using E_Commerce.DataAccess.Constants;
 
 namespace E_Commerce.Web.Areas.Api.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class ProductController : ControllerBase
     {
         private readonly IProductService _productService;
@@ -21,6 +22,8 @@ namespace E_Commerce.Web.Areas.Api.Controllers
         }
 
         [HttpGet("All")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<IEnumerable<ProductDto>>> GetAll()
         {
             var products = await _productService.GetAllProductsDtoAsync();
@@ -30,6 +33,8 @@ namespace E_Commerce.Web.Areas.Api.Controllers
         }
 
         [HttpGet("{id}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<ProductDto>> Get(int id)
         {
             if (id <= 0)
@@ -41,15 +46,22 @@ namespace E_Commerce.Web.Areas.Api.Controllers
         }
 
         [HttpPost("Add")]
+        [Authorize(Roles = AppRoles.Admin)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<ActionResult<ProductDto>> AddProduct([FromForm] ProductPostDto dto)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
             var created = await _productService.AddProductAsync(dto);
-            return CreatedAtAction(nameof(Get), new { id = created.Id }, created);
+            return Created();
         }
 
         [HttpPut("Update/{id}")]
+        [Authorize(Roles = AppRoles.Admin)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<ActionResult<ProductDto>> UpdateProduct(int id, [FromBody] ProductUpdateDto dto)
         {
             if (id != dto.Id)
@@ -61,8 +73,13 @@ namespace E_Commerce.Web.Areas.Api.Controllers
         }
 
         [HttpDelete("Delete/{id}")]
+        [Authorize(Roles = AppRoles.Admin)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<IActionResult> DeleteProduct(int id)
         {
+            if (id <= 0) return BadRequest();
             var deleted = await _productService.DeleteProductAsync(id);
             if (!deleted)
                 return NotFound();
